refactor: put everything in a new file
